feat(jzero): support jzero template

+++ embeded/template.go
@@ -2,18 +2,30 @@ package embeded
 
 import (
 	"embed"
-	"log"
 	"os"
 	"path/filepath"
+
+	"github.com/zeromicro/go-zero/tools/goctl/util/pathx"
 )
 
-var Template embed.FS
+var (
+	Template embed.FS
+	// Home template home
+	Home string
+)
 
 func ReadTemplateFile(filename string) []byte {
 	path := filepath.ToSlash(filepath.Join(".template", filename))
 	data, err := Template.ReadFile(path)
 	if err != nil {
 		return nil
+	}
+	if Home != "" {
+		file, err := os.ReadFile(filepath.Join(Home, filename))
+		if err != nil {
+			return data
+		}
+		return file
 	}
 	return data
 }
@@ -26,7 +38,7 @@ func WriteTemplateDir(sourceDir, targetDir string) error {
 
 	err = writeTemplateDirRecursive(sourceDir, targetDir)
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	return nil
@@ -53,9 +65,18 @@ func writeTemplateDirRecursive(sourceDir, targetDir string) error {
 				return err
 			}
 		} else {
+			if pathx.FileExists(targetPath) {
+				continue
+			}
+
 			data, err := Template.ReadFile(filepath.ToSlash(filepath.Join(".template", sourcePath)))
 			if err != nil {
 				return err
+			}
+
+			file, err := os.ReadFile(filepath.Join(Home, sourcePath))
+			if err == nil {
+				data = file
 			}
 
 			err = os.WriteFile(targetPath, data, 0o644)
